refactor: moved useProfileGivenAddress into hooks/profile
